# Fix and optimize complex identifier parsing