… src/Esprima/Character.cs