Fixes some bugs (mostly related to handling of surrogate characters) which I found around complex identifiers while working on regexp parsing. (Kinda a fix for this earlier [fix](https://github.com/sebastienros/esprima-dotnet/pull/379).)

Also reduces memory allocation in `Scanner.GetComplexIdentifier` and improves performance a tiny bit:

## Esprima.Benchmark.FileParsingBenchmark

| **Diff**|Method|FileName|Mean|Error|Allocated|
|------- |-------|-------|-------:|-------|-------:|
| Old |ParseProgram|angular-1.2.5|8.610 ms|0.0455 ms|4293.9 KB|
| **New** |	|	| **8,592.6 μs (-3%)** | **39.54 μs** | **4293.9 KB (0%)** |
| Old |ParseProgram|backbone-1.1.0|1.142 ms|0.0050 ms|673.02 KB|
| **New** |	|	| **1,154.0 μs (-1%)** | **4.76 μs** | **673.02 KB (0%)** |
| Old |ParseProgram|jquery-1.9.1|6.764 ms|0.0216 ms|3744.63 KB|
| **New** |	|	| **6,671.5 μs (-4%)** | **16.01 μs** | **3744.63 KB (0%)** |
| Old |ParseProgram|jquery.mobile-1.4.2|10.289 ms|0.0189 ms|5762.71 KB|
| **New** |	|	| **10,311.4 μs (-2%)** | **27.14 μs** | **5762.71 KB (0%)** |
| Old |ParseProgram|mootools-1.4.5|5.445 ms|0.0212 ms|3074.04 KB|
| **New** |	|	| **5,367.2 μs (-4%)** | **23.21 μs** | **3074.04 KB (0%)** |
| Old |ParseProgram|underscore-1.5.2|1.033 ms|0.0491 ms|568.59 KB|
| **New** |	|	| **957.7 μs (-9%)** | **8.92 μs** | **568.59 KB (0%)** |
| Old |ParseProgram|yui-3.12.0|5.093 ms|0.0828 ms|2827.59 KB|
| **New** |	|	| **4,789.2 μs (-8%)** | **20.64 μs** | **2827.59 KB (0%)** |